RF_01.001.10 | New Item > Creatе a new item > Create Freestyle project
