+++ src/test/java/school/redrover/MultibranchPipelineTest.java
@@ -40,78 +40,4 @@ public class MultibranchPipelineTest extends BaseTest {
         Assert.assertEquals(actualDescription, PROJECT_DESCRIPTION);
     }
 
-    @Test
-    public void testCreateWithSpecialSymbols() {
-        String[] specialCharacters = {"!", "%", "&", "#", "@", "*", "$", "?", "^", "|", "/", "]", "["};
-
-        getDriver().findElement(By.cssSelector("[href$='/newJob']")).click();
-
-        JavascriptExecutor js = (JavascriptExecutor) getDriver();
-        js.executeScript("window.scrollTo(0, document.body.scrollHeight);");
-
-        getDriver().findElement(By.cssSelector("[class$='MultiBranchProject']")).click();
-        WebElement nameField = getDriver().findElement(By.xpath("//input[@name='name']"));
-
-        for (String specChar : specialCharacters) {
-            nameField.clear();
-            nameField.sendKeys("Mult" + specChar + "branch");
-
-            WebElement actualMessage = getWait10().until(ExpectedConditions.visibilityOfElementLocated(By.
-                    xpath("//div[@id='itemname-invalid']")));
-
-            String expectMessage = "» ‘" + specChar + "’ is an unsafe character";
-            Assert.assertEquals(actualMessage.getText(), expectMessage, "Message is not displayed");
-        }
-    }
-
-    @Test
-    public void testVerifySectionHasTooltip() {
-        getDriver().findElement(By.cssSelector("[href$='/newJob']")).click();
-        getDriver().findElement(By.id("name")).sendKeys("MultiBranchPipeline");
-
-        JavascriptExecutor js = (JavascriptExecutor) getDriver();
-        js.executeScript("window.scrollTo(0, document.body.scrollHeight);");
-
-        getDriver().findElement(By.cssSelector("[class$='MultiBranchProject']")).click();
-        getDriver().findElement(By.id("ok-button")).click();
-
-        String actualTooltip = getDriver().
-                findElement(By.xpath("//*[contains(text(), '" + "Display Name" + "')]//a"))
-                .getAttribute("tooltip");
-
-        Assert.assertEquals(actualTooltip, "Help for feature: Display Name");
-    }
-
-    @Ignore
-    @Test
-    public void testTryCreateProjectExistName() throws InterruptedException {
-
-        final String projectName = "MultiBuild";
-        final String errorMessage = "» A job already exists with the name " + "‘" + projectName + "’";
-
-        getWait10().until(ExpectedConditions.visibilityOfElementLocated(By.xpath("//*[@id='tasks']/div[1]/span/a"))).click();
-        getDriver().findElement(By.className("jenkins-input")).sendKeys(projectName);
-
-        JavascriptExecutor js = (JavascriptExecutor) getDriver();
-        js.executeScript("window.scrollTo(0, document.body.scrollHeight);");
-
-        getDriver().findElement(By.cssSelector("[class$='MultiBranchProject']")).click();
-        getDriver().findElement(By.id("ok-button")).click();
-
-        JavascriptExecutor js1 = (JavascriptExecutor) getDriver();
-        js1.executeScript("window.scrollTo(0, document.body.scrollHeight);");
-
-        getDriver().findElement(By.name("Submit")).click();
-        getWait10().until(ExpectedConditions.visibilityOfElementLocated(By.xpath("//*[@id='jenkins-name-icon']"))).click();
-
-        getWait10().until(ExpectedConditions.visibilityOfElementLocated(By.xpath("//*[@id='tasks']/div[1]/span/a"))).click();
-        JavascriptExecutor js2 = (JavascriptExecutor) getDriver();
-        js2.executeScript("window.scrollTo(0, document.body.scrollHeight);");
-
-        getDriver().findElement(By.id("name")).sendKeys(projectName);
-        getDriver().findElement(By.cssSelector("[class$='MultiBranchProject']")).click();
-
-        String actualMessage = getDriver().findElement(By.xpath("//*[@id='itemname-invalid']")).getText();
-        Assert.assertEquals(actualMessage, errorMessage);
-    }
 }
